+++ LateForDinner/Assets/Scripts/UserInterfaces/Displays/UISplashDisplay.cs
@@ -36,7 +36,7 @@ public class UISplashDisplay : UIDisplay, IAnimatable
 
     private void SetAlpha(Images imageType, float alpha)
     {
-        var image = GetImage((int)imageType);
+        var image = GetImage(imageType);
 
         if (image == null)
             return;
@@ -51,9 +51,9 @@ public class UISplashDisplay : UIDisplay, IAnimatable
         try
         {
             var token = GetToken("SplashTask");
-            var unityImage = GetImage((int)Images.UnityImage);
-            var teamImage = GetImage((int)Images.TeamImage);
-            var titleImage = GetImage((int)Images.TitleImage);
+            var unityImage = GetImage(Images.UnityImage);
+            var teamImage = GetImage(Images.TeamImage);
+            var titleImage = GetImage(Images.TitleImage);
             await unityImage.FadeAsync(0f, 1f, 1.2f, token: token);
             await UniTask.Delay(1000, cancellationToken: token);
             await unityImage.FadeAsync(1f, 0f, 1.2f, token: token);
@@ -61,7 +61,7 @@ public class UISplashDisplay : UIDisplay, IAnimatable
             await UniTask.Delay(1000, cancellationToken: token);
             await teamImage.FadeAsync(1f, 0f, 1.2f, token: token);
             await UniTask.Delay(1000, cancellationToken: token);
-            GetImage((int)Images.BackgroundImage).color = new Color(1f, 1f, 1f, 0f);
+            GetImage(Images.BackgroundImage).color = new Color(1f, 1f, 1f, 0f);
             await titleImage.FadeAsync(0f, 1f, 0.2f, 2.0f, token: token);
             await UniTask.Delay(2000, cancellationToken: token);
         }
